feat(#701): remove DirectiveAnnotationProperty class at all
